Files without 'end sector' now are detected and marked with 'LOOP'
A file which has no last sector - sector pointing to sector #0 - will be marked with 'LOOP'
This could happen when a file was deleted and one or more of it's sectors where  overwritten.

Older versions of 'atdir' were caugth in an infinity loop when attempting to read such a
files sector chain.